fix: remove non Dutch labels and definitions from NUTS code list
The NUTS LAU 2024 code list was initially imported in full in #34. The presence
of language tagged `skos:prefLabel`s causes the frontend to show duplicate
options in the relevant drop down lists.
This modifies the initial migration by deleting the English language tagged
triples. On environments where the migration was already executed (DEV and TEST)
a local migration to delete the unnecessary triples should be executed.