fix: add suffix to labels for pre 2024 merger NUTS codes
The 2024 NUTS code list uses the same label for codes when a municipality's
name does **not** change as part of the 2025 mergers. This results in seemingly
duplicate options when a user wants to select a product's scope (nl. Geografisch
toepassingsgebied).

This adds a suffix to the labels for the NUTS codes of the "old" municipalities.
This way users to select the right scope based on the label. Note, in to
function correctly LPDC needs to have both the "old" and "new" NUTS codes for
the municipalities, so removing the "old" is not an option.