chore: example migrations for product 2025 merger instance archival
DO NOT MERGE: this is only meant as an example to simplify reviewing. The
actual migrations execution will be done locally on each individual environment.

Example migrations to archive product instances after the 2025 municipality
mergers. There were generated using the `archive-instances-script.ts` defined as
part of the `lpdc-management-service`.

These migrations
- remove product instances from LPDC itself; and
- tombstone published product instance such that their archival is also
  published to IPDC.