fix: link organisations to NUTS2024 spatials
As part of the 2025 municipality mergers several spatials became invalid as
their municipalities were involved in a merger. For active organisations we
update their werkingsgebieden to link to the new, unexpired spatials. This way
the correct, unexpired spatial is added as geographical area when those
organsations create new product instances.