chore: bump LDES consumer for Ghent
Upgrade the LDES consumer for Ghent's instance snapshot feed to the new version
which is actively maintained. Furthermore, this version correctly supports
obtaining the first node from the viewpoint instead of hardcoding it in the
viewpoint URL.This allows Ghent to apply a retention policy which may cause the
page number of the first page to vary over time.